refactor/ fallback 네이밍 통일
