+++ src/app/(home)/_components/ContentSliderFallback.tsx
